Added Application Logic Added mapper

+++ src/PrintRelayServer.Shared/Contracts/Base/PaginatedResult.cs
@@ -2,9 +2,9 @@
 
 public class PaginatedResult<TEntity>
 {
-    public uint TotalCount { get; set; } = 0;
-    public uint PageNumber { get; set; }
-    public uint PageSize { get; set; }
-    public uint TotalPages { get; set; } = 0;
+    public long TotalCount { get; set; } = 0;
+    public int PageNumber { get; set; }
+    public int PageSize { get; set; }
+    public int TotalPages { get; set; } = 0;
     public IList<TEntity> Results { get; set; } = new List<TEntity>();
 }
